+++ scripts/generate-fixtures.cjs
@@ -1,0 +1,7 @@
+const path = require("path");
+
+// Import all projection generators
+require("./fixtures/projections/gnomonic.cjs");
+require("./fixtures/projections/authalic.cjs");
+
+console.log("All projection fixtures generated successfully!");
